fix: findByIdAndUpdate, return in if statement, delete /posts/ error

+++ controller/posts.js
@@ -13,15 +13,15 @@ export default {
     try {
       const { content, ...otherData } = req.body
       if (!content) {
-        errorHandler({ res, statusCode: 400, message: '內容不得為空' })
+        return errorHandler({ res, statusCode: 400, message: '內容不得為空' })
       }
 
       const postRes = await Post.create({ content, ...otherData })
       console.log('postRes', postRes)
 
-      successHandler({ res, data: postRes })
+      return successHandler({ res, data: postRes })
     } catch (error) {
-      errorHandler({ res, statusCode: 400, error: error.message })
+      return errorHandler({ res, statusCode: 400, error: error.message })
     }
   },
 
@@ -30,12 +30,20 @@ export default {
     try {
       const { content, ...otherData } = req.body
       if (!content) {
-        errorHandler({ res, statusCode: 400, message: 'content is required' })
+        return errorHandler({
+          res,
+          statusCode: 400,
+          message: 'content is required',
+        })
       }
 
       const id = req.params.post_id
 
-      const post = await Post.findByIdAndUpdate(id, { content, ...otherData }, { returnDocument: 'after' })
+      const post = await Post.findByIdAndUpdate(
+        id,
+        { content, ...otherData },
+        { returnDocument: 'after', runValidators: true },
+      )
       if (!id || !post) {
         return errorHandler({
           res,
@@ -47,13 +55,16 @@ export default {
       return successHandler({ res, data: post })
     } catch (error) {
       return errorHandler({
-        res, statusCode: 400, error: error.message,
+        res,
+        statusCode: 400,
+        error: error.message,
       })
     }
   },
 
   // DELETE /posts/:id
   async deletePost(req, res) {
+    console.log('123')
     try {
       const id = req.params.post_id
       const data = await Post.findByIdAndDelete(id)
@@ -67,19 +78,27 @@ export default {
       }
       return successHandler({ res, data, message: '刪除成功' })
     } catch (error) {
-      return errorHandler({ res, code: 400, errorMessage: error.message })
+      return errorHandler({ res, code: 400, message: error.message })
     }
   },
 
   // DELETE /posts
   async deletePosts(req, res) {
+    console.log('req.originalUrl', req.originalUrl)
+    console.log(req.originalUrl === '/api/posts/')
+    if (req.originalUrl === '/api/posts/') {
+      return errorHandler({
+        res,
+        statusCode: 400,
+        message: 'not allowed api path',
+      })
+    }
     try {
       const deleteRes = await Post.deleteMany()
       const posts = await Post.find()
       return successHandler({ res, data: posts, message: '刪除成功' })
     } catch (error) {
-      return errorHandler({ res, code: 400, errorMessage: error.message })
+      return errorHandler({ res, statusCode: 400, message: error.message })
     }
   },
-
 }
